requirements.txt
streamlit==1.30.0
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
requests==2.31.0
numpy==1.26.3